requirements.txt
absl-py==1.4.0
attrs==23.1.0
cffi==1.15.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
flatbuffers==23.5.26
fonttools==4.40.0
kiwisolver==1.4.4
matplotlib==3.7.1
mediapipe==0.10.1
mutagen==1.46.0
numpy==1.24.3
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sounddevice==0.4.6
